ai_service/requirements.txt
fastapi
uvicorn
pytesseract
pillow
PyMuPDF
openai
python-multipart
python-dotenv
